# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Translation Management API",
        "description": "API for managing translations in multiple languages",
        "contact": {
            "email": "haider8278@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Create a Auth Token",
                "operationId": "c5e838284c7f343f0e5e5fd589161c96",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Translation created"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": []
            }
        },
        "/api/translations": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Retrieve a list of translations",
                "description": "Returns a paginated list of translations with optional filters.",
                "operationId": "434a55804ca50c9845bd4781c222a539",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "query",
                        "description": "Filter by locale (e.g., en, fr, es)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Filter by translation key",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "0": "sanctum"
                    }
                ]
            },
            "post": {
                "tags": [
                    "Translations"
                ],
                "summary": "Create a new translation",
                "operationId": "fe38d8cf02ff5c92d2cd8d5dfb095e5f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "locale",
                                    "key",
                                    "content"
                                ],
                                "properties": {
                                    "locale": {
                                        "type": "string",
                                        "example": "en"
                                    },
                                    "key": {
                                        "type": "string",
                                        "example": "welcome_message"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Welcome!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Translation created"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/translations/{id}": {
            "put": {
                "tags": [
                    "Translations"
                ],
                "summary": "Update an existing translation",
                "operationId": "e314c7afaf04580ad65b747c8a1f7ce6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Translation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Hello, world!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Translation updated"
                    },
                    "404": {
                        "description": "Translation not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Translations"
                ],
                "summary": "Delete a translation",
                "operationId": "7d0107320d96a296dcfcb666934c0e3a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Translation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Translation deleted"
                    },
                    "404": {
                        "description": "Translation not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/translations/json-export": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Retrieve all translations in JSON format",
                "description": "Returns all translations optimized with Redis caching.",
                "operationId": "68c6a3d08d78c1d71bc65aff5c625543",
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Endpoints for Login and auth token"
        },
        {
            "name": "Translations",
            "description": "Endpoints for managing translations"
        }
    ],
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "query"
            }
        }
    }
}